requirements.txt
opencv-python
Pillow
python-dotenv
google-generativeai
pyttsx3
SpeechRecognition 
pyaudio